Address review comments.
Move all the implementation to the .h file.
Move uArchInterfaces to uArchBase.
