+++ mlir/lib/Dialect/LLVMIR/CMakeLists.txt
@@ -129,6 +129,5 @@ add_mlir_dialect_library(MLIRXeVMDialect
   MLIRDialectUtils
   MLIRIR
   MLIRLLVMDialect
-  MLIRXeGPUuArch
   MLIRSideEffectInterfaces
 )
