+++ mlir/lib/Dialect/XeGPU/CMakeLists.txt
@@ -1,4 +1,3 @@
 add_subdirectory(IR)
 add_subdirectory(Transforms)
-add_subdirectory(uArch)
 add_subdirectory(Utils)
