+++ mlir/lib/Dialect/XeGPU/IR/CMakeLists.txt
@@ -20,7 +20,6 @@ add_mlir_dialect_library(MLIRXeGPUDialect
   MLIRGPUDialect
   MLIRXeVMDialect
   MLIRIR
-  MLIRXeGPUuArch
   MLIRViewLikeInterface
   MLIRVectorDialect
 )
