+++ mlir/lib/Dialect/XeGPU/Transforms/CMakeLists.txt
@@ -23,7 +23,6 @@ add_mlir_dialect_library(MLIRXeGPUTransforms
   MLIRTransforms
   MLIRGPUDialect
   MLIRXeGPUUtils
-  MLIRXeGPUuArch
   MLIRGPUUtils
   MLIRVectorTransforms
 )
