--- mlir/lib/Dialect/XeGPU/uArch/IntelGpuXe2.cpp
@@ -1,172 +1,0 @@
-#include "mlir/Dialect/XeGPU/uArch/IntelGpuXe2.h"
-#include "mlir/IR/BuiltinTypes.h"
-#include "llvm/Support/DebugLog.h"
-#include <algorithm>
-#include <vector>
-
-#define DEBUG_TYPE "xegpu-uarch"
-
-using namespace mlir::xegpu::uArch;
-using namespace mlir::xegpu::uArch::Xe2Plus;
-
-namespace mlir {
-namespace xegpu {
-namespace uArch {
-namespace Xe2Plus {
-
-std::vector<std::pair<uint32_t, uint32_t>>
-DPASInstruction::getSupportedShapes(mlir::Type dataType,
-                                    MMAOpndKind matrixType) {
-  auto combineVectors = [](const std::vector<uint32_t> &a,
-                           const std::vector<uint32_t> &b)
-      -> std::vector<std::pair<uint32_t, uint32_t>> {
-    std::vector<std::pair<uint32_t, uint32_t>> result;
-    for (unsigned x : a) {
-      for (unsigned y : b) {
-        result.emplace_back(x, y);
-      }
-    }
-    return result;
-  };
-
-  auto M = getSupportedM(dataType);
-  auto K = getSupportedK(dataType);
-  auto N = getSupportedN(dataType);
-  std::vector<std::pair<unsigned, unsigned>> resultMatrix;
-
-  switch (matrixType) {
-  case MMAOpndKind::MatrixA:
-    resultMatrix = combineVectors(M, K);
-    break;
-  case MMAOpndKind::MatrixB:
-    resultMatrix = combineVectors(K, N);
-    break;
-  case MMAOpndKind::MatrixC:
-    resultMatrix = combineVectors(M, N);
-    break;
-  case MMAOpndKind::MatrixD:
-    resultMatrix = combineVectors(M, N);
-    break;
-  }
-  return resultMatrix;
-}
-
-std::vector<mlir::Type>
-DPASInstruction::getSupportedTypes(MLIRContext &context,
-                                   MMAOpndKind matrixType) {
-  mlir::Type bf16Type = mlir::BFloat16Type::get(&context);
-  mlir::Type f16Type = mlir::Float16Type::get(&context);
-  mlir::Type tf32Type = mlir::FloatTF32Type::get(&context);
-  mlir::Type f32Type = mlir::Float32Type::get(&context);
-
-  switch (matrixType) {
-  case MMAOpndKind::MatrixA:
-    return {bf16Type, f16Type, tf32Type};
-    break;
-  case MMAOpndKind::MatrixB:
-    return {bf16Type, f16Type, tf32Type};
-    break;
-  case MMAOpndKind::MatrixC:
-    return {bf16Type, f16Type, f32Type};
-    break;
-  case MMAOpndKind::MatrixD:
-    return {bf16Type, f16Type, f32Type};
-    break;
-  }
-}
-
-bool DPASInstruction::checkSupportedTypes(mlir::Type AType, mlir::Type BType,
-                                          mlir::Type CType, mlir::Type DType) {
-  if (AType.isF16() || BType.isF16()) {
-    if (AType != BType || (CType && (!CType.isF32() && !CType.isF16())) ||
-        (!DType.isF32() && !DType.isF16())) {
-      LDBG() << "Unsupported dpas combinations of Dst, Acc, A and B matrices.";
-      return false;
-    }
-  } else if (AType.isBF16() || BType.isBF16()) {
-    if (AType != BType || (CType && (!CType.isF32() && !CType.isBF16())) ||
-        (!DType.isF32() && !DType.isBF16())) {
-      LDBG() << "Unsupported dpas combinations of Dst, Acc, A and B matrices.";
-      return false;
-    }
-  } else if (AType.isTF32() || BType.isTF32()) {
-    if (AType != BType || (CType && (!CType.isF32() && !DType.isF32())) ||
-        (!DType.isF32())) {
-      LDBG() << "Unsupported dpas combinations of Dst, Acc, A and B matrices.";
-      return false;
-    }
-  } else if (!(AType.isInteger(2) || AType.isInteger(4) ||
-               AType.isInteger(8)) &&
-             !(BType.isInteger(2) || BType.isInteger(4) ||
-               BType.isInteger(8))) {
-    LDBG() << "Unsupported dpas combinations of Dst, Acc, A and B matrices.";
-    return false;
-  }
-
-  return true;
-}
-
-bool DPASInstruction::checkSupportedShapesAndTypes(
-    std::pair<uint32_t, uint32_t> AShape, std::pair<uint32_t, uint32_t> BShape,
-    std::pair<uint32_t, uint32_t> CShape, std::pair<uint32_t, uint32_t> DShape,
-    mlir::Type AType, mlir::Type BType, mlir::Type CType, mlir::Type DType) {
-  auto supportedAShapes = getSupportedShapes(AType, MMAOpndKind::MatrixA);
-  auto supportedBShapes = getSupportedShapes(BType, MMAOpndKind::MatrixB);
-  auto supportedCShapes = getSupportedShapes(CType, MMAOpndKind::MatrixC);
-  auto supportedDShapes = getSupportedShapes(DType, MMAOpndKind::MatrixD);
-  return llvm::is_contained(supportedAShapes, AShape) &&
-         llvm::is_contained(supportedBShapes, BShape) &&
-         llvm::is_contained(supportedCShapes, CShape) &&
-         llvm::is_contained(supportedDShapes, DShape) &&
-         checkSupportedTypes(AType, BType, CType, DType);
-}
-
-bool DPASInstruction::validate(std::pair<uint32_t, uint32_t> AShape,
-                               std::pair<uint32_t, uint32_t> BShape,
-                               std::pair<uint32_t, uint32_t> CShape,
-                               std::pair<uint32_t, uint32_t> DShape,
-                               mlir::Type AType, mlir::Type BType,
-                               mlir::Type CType, mlir::Type DType) {
-  return checkSupportedShapesAndTypes(AShape, BShape, CShape, DShape, AType,
-                                      BType, CType, DType);
-}
-
-std::vector<uint32_t> DPASInstruction::getSupportedM(mlir::Type type) {
-  return {1, 2, 3, 4, 5, 6, 7, 8};
-}
-
-std::vector<uint32_t> DPASInstruction::getSupportedK(mlir::Type type) {
-  // assert if data type is not int or float type
-  assert(type.isIntOrFloat() && "Matrix type must be int or float");
-  auto bitWidth = type.getIntOrFloatBitWidth();
-  uint32_t kSize = 0;
-  switch (bitWidth) {
-  case 2:
-    kSize = 64;
-    break;
-  case 4:
-    kSize = 64;
-    break;
-  case 8:
-    kSize = 32;
-    break;
-  case 16:
-    kSize = 16;
-    break;
-  case 32:
-    kSize = 8;
-    break;
-  default:
-    llvm_unreachable("Invalid int or float");
-  }
-  return {kSize};
-}
-
-std::vector<uint32_t> DPASInstruction::getSupportedN(mlir::Type type) {
-  return {16};
-}
-
-} // namespace Xe2Plus
-} // namespace uArch
-} // namespace xegpu
-} // namespace mlir
